move alternatives to separate files
